requirements.txt
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==2.0.0
pytesseract==0.3.10
Pillow==10.0.0
opencv-python==4.8.0.74
spacy==3.6.0
nltk==3.8.1
flask==2.3.3
flask-cors==4.0.0
joblib==1.3.1
requests==2.31.0
